+++ src/tx-params/redeem-check.js
@@ -9,9 +9,10 @@ import {toBuffer} from 'minterjs-util';
 
 
 /**
+ * //@TODO https://github.com/MinterTeam/minter-js-sdk/issues/13 to allow easy `prepareLink` without proof
  * @constructor
  * @param {string|Buffer} privateKey
- * @param {string} check
+ * @param {string|Buffer} check
  * @param {string} [password]
  * @param {string|Buffer} [proof]
  * @param {string} [feeCoinSymbol] - should be base coin